+++ include/day2.h
@@ -6,8 +6,8 @@ extern "C" {
 
 #include <stddef.h>
 
-int day2_part1(const char** lines, size_t nlines);
-int day2_part2(const char** lines, size_t nlines);
+long day2_part1(const char** lines, size_t nlines);
+long day2_part2(const char** lines, size_t nlines);
 
 #ifdef __cplusplus
 }
