+++ include/day3.h
@@ -6,8 +6,8 @@ extern "C" {
 
 #include <stddef.h>
 
-int day3_part1(const char** lines, size_t nlines);
-int day3_part2(const char** lines, size_t nlines);
+long day3_part1(const char** lines, size_t nlines);
+long day3_part2(const char** lines, size_t nlines);
 
 #ifdef __cplusplus
 }
