+++ include/day4.h
@@ -6,8 +6,8 @@ extern "C" {
 
 #include <stddef.h>
 
-int day4_part1(const char** lines, size_t nlines);
-int day4_part2(const char** lines, size_t nlines);
+long day4_part1(const char** lines, size_t nlines);
+long day4_part2(const char** lines, size_t nlines);
 
 #ifdef __cplusplus
 }
