+++ include/day5.h
@@ -6,8 +6,8 @@ extern "C" {
 
 #include <stddef.h>
 
-int day5_part1(const char** lines, size_t nlines);
-int day5_part2(const char** lines, size_t nlines);
+long day5_part1(const char** lines, size_t nlines);
+long day5_part2(const char** lines, size_t nlines);
 
 #ifdef __cplusplus
 }
