+++ include/day7.h
@@ -6,8 +6,8 @@ extern "C" {
 
 #include <stddef.h>
 
-int day7_part1(const char** lines, size_t nlines);
-int day7_part2(const char** lines, size_t nlines);
+long day7_part1(const char** lines, size_t nlines);
+long day7_part2(const char** lines, size_t nlines);
 
 #ifdef __cplusplus
 }
