+++ include/day8.h
@@ -6,8 +6,8 @@ extern "C" {
 
 #include <stddef.h>
 
-int day8_part1(const char** lines, size_t nlines);
-int day8_part2(const char** lines, size_t nlines);
+long day8_part1(const char** lines, size_t nlines);
+long day8_part2(const char** lines, size_t nlines);
 
 #ifdef __cplusplus
 }
