+++ src/main.c
@@ -17,7 +17,7 @@ __attribute__((noreturn)) void usage_abrt(char* progname) {
   exit(EX_USAGE);
 }
 
-typedef int (*solver_t)(const char**, size_t);
+typedef long (*solver_t)(const char**, size_t);
 static solver_t solvers[][2] = {
     {day1_part1, day1_part2}, {day2_part1, day2_part2},
     {day3_part1, day3_part2}, {day4_part1, day4_part2},
@@ -42,7 +42,7 @@ int main(int argc, char** argv) {
   size_t nlines;
   const char** lines = readlines(argv[3], &nlines);
 
-  printf("Result: %d\n", solver(lines, nlines));
+  printf("Result: %ld\n", solver(lines, nlines));
 
   freelines(lines);
 }
